enforce usage of uint32 based numbers rather than int32, so that they are always positive, cleaner behavior.

+++ protocol/common.go
@@ -4,8 +4,8 @@ import (
 	"github.com/google/uuid"
 )
 
-func GenerateId() int32 {
-	return int32(uuid.New().ID())
+func GenerateId() uint32 {
+	return uuid.New().ID()
 }
 
 func AppendUnique[T comparable](s []T, el T) []T {
